Network.HTTP.ReverseProxy

Types

data ProxyDest

Host/port combination to which we want to proxy.

Constructors

ProxyDest

Fields

pdHost :: !ByteString

pdPort :: !Int

Raw

rawProxyTo

Arguments

:: (MonadBaseControl IO m, MonadIO m)

=> (RequestHeaders -> m (Either (Application m) ProxyDest))
How to reverse proxy. A Left result will run the given Application, whereas a Right will reverse proxy to the given host/port.

-> Application m

Set up a reverse proxy server, which will have a minimal overhead.

This function uses raw sockets, parsing as little of the request as possible. The workflow is:

Parse the first request headers.

Ask the supplied function to specify how to reverse proxy.

Open up a connection to the given host/port.

Pass all bytes across the wire unchanged.

If you need more control, such as modifying the request or response, use waiProxyTo.

WAI + http-conduit

waiProxyTo

Arguments

:: (Request -> ResourceT IO (Either Response ProxyDest))
How to reverse proxy. A Left result will be sent verbatim as the response, whereas Right will cause a reverse proxy.

-> (SomeException -> Application)
How to handle exceptions when calling remote server. For a simple 502 error page, use defaultOnExc.

-> Manager
connection manager to utilize

-> Application

Creates a WAI Application which will handle reverse proxies.

Connections to the proxied server will be provided via http-conduit. As such, all requests and responses will be fully processed in your reverse proxy. This allows you much more control over the data sent over the wire, but also incurs overhead. For a lower-overhead approach, consider rawProxyTo.

Most likely, the given application should be run with Warp, though in theory other WAI handlers will work as well.

Note: This function will use chunked request bodies for communicating with the proxied server. Not all servers necessarily support chunked request bodies, so please confirm that yours does (Warp, Snap, and Happstack, for example, do).

defaultOnExc :: SomeException -> Application

Sends a simple 502 bad gateway error message with the contents of the exception.

WAI to Raw

waiToRaw :: Application -> Application IO

Convert a WAI application into a raw application, using Warp.